refactor go binding module/client to use struct
Problem: we currently pass around a ctx to interact with all
go binding module or client functions, and provide the developer
user with context that we are using the clients via function
names (e.g., ReapiCliMatchAllocate).
Solution: we can save the context alongside a struct that is
returned with XNew(), and then rename the functions to not
include the extra information (e.g., cli.MatchAllocate() implies
MatchAllocate for the ReapiClient{ctx:ctx} that it is created
from.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>